Attempt to fix Nethermind permissions.
The Nethermind chiseled dockerfile uses a non-root user and it sets up some very specific paths with proper permissions.  We were ignoring the default paths with permissions setup and using our own (so we could have a single volume) but this breaks the permission setup.  Docker, sadly, doesn't have a way to fix this at runtime without manual intervention, so this change switches to using Nethermind supplied paths.